+++ src/main/java/alkalus/main/mixins/early/minecraft/NetHandlerPlayClientMixin.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.injection.At;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 
-import alkalus.main.core.util.ClientRenderCache;
+import alkalus.main.mixins.hooks.ClientRenderCache;
 
 @Mixin(NetHandlerPlayClient.class)
 public class NetHandlerPlayClientMixin {
